move validation code to validate_domain.go
